# internal/docs/swagger.yaml
basePath: /v1
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact:
    email: info@stormlightlabs.org
    name: API Support
    url: https://github.com/stormlightlabs/baseball
  description: |
    <!-- markdownlint-disable -->

    A comprehensive REST API for baseball statistics serving data from the Lahman Baseball Database and Retrosheet

    ## Attributions

    **[SABR Lahman Database](https://sabr.org/lahman-database/)**: The information used here was obtained free of charge from and is copyrighted by Sean Lahman.

    **[Retrosheet.org](https://www.retrosheet.org/)**: The information used here was obtained free of charge from and is copyrighted by Retrosheet.

    wOBA weights and seasonal constants from **[FanGraphs Guts!](https://www.fangraphs.com/tools/guts)**, used under FanGraphs’ published terms.
  license:
    name: MPL-2.0
    url: https://opensource.org/license/mpl-2-0
  title: Baseball API
  version: "1.0"
paths:
...
  /search/games:
    get:
      consumes:
      - application/json
      description: Natural language game search supporting queries like "yankees red
        sox 2004 alcs game 7" or "dodgers giants 2014 nlcs"
      parameters:
      - description: Natural language search query
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Results per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Search games with natural language
      tags:
      - search
      - games
  /search/parks:
    get:
      consumes:
      - application/json
      description: Ballpark lookup by name, city, state, or park ID
      parameters:
      - description: Search query (searches park name, city, state, and park ID)
        in: query
        name: q
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Results per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Search parks
      tags:
      - search
      - parks
  /search/players:
    get:
      consumes:
      - application/json
      description: Fuzzy player search with filters for name, position, era, and more
      parameters:
      - description: Search query (searches first and last name)
        in: query
        name: q
        type: string
      - description: Filter by debut year
        in: query
        name: debut_year
        type: integer
      - description: Filter by position
        in: query
        name: position
        type: string
      - description: Filter by batting hand (R, L, B)
        in: query
        name: bats
        type: string
      - description: Filter by throwing hand (R, L)
        in: query
        name: throws
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Results per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Search players
      tags:
      - search
      - players
  /search/teams:
    get:
      consumes:
      - application/json
      description: Search teams by name, city, or franchise
      parameters:
      - description: Search query (searches team name, team ID, and franchise ID)
        in: query
        name: q
        type: string
      - description: Filter by season year
        in: query
        name: year
        type: integer
      - description: Filter by league (AL, NL)
        in: query
        name: league
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Results per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Search teams
      tags:
      - search
      - teams
...
swagger: "2.0"
...